Add lambda to scheme interpreter
This is a large commit that contains a lot of different pieces of
code, but this is the one that is finished. This adds a lambda
construct, which allows you to create anonymous functions, as seen
below

((lambda (x) (* x 2) ; Creates an anonymous method that doubles a number
 3) ; passes in 3 to that method

You can use define to save this in a variable:

(define id (lambda (x) (* x 2)))

This is just a more verbose and consistent way of writing this:

(define (id x) (* x 2))

This code change required rewriting the way that the lisp interpreter
handles methods, which should be helpful later.

A quick overview of some of the other code that went into this commit:
- Make memory.c work on -O2, which previously wouldn't work
- Add warnings and -Werror to the default compilation options, and
cleaned up the code accordingly
- Initial support for a tar(1) based ramdisk, which is constructed
automatically
- Make halt disable interrupts, so multithreaded code will also stop.
- Make some variable names more consistent (pascal_case)
- Allocation tracking, which will be used when I add garbage
collection to the interpreter.
- Make the interrupt handler code work cleaner
- Refactor scope initialization to avoid weird bugs with the temporary variable.
- Fix a buffer overflow in the scope_add function
- Update some parts of the code that use the name scope2. This is
vestigal from when I was updating the scope implementation to use a hashmap.

+++ src/keyboard.c
@@ -168,24 +168,24 @@ i8 keyboard_scan() {
 	if (ret >= 0x80) {
 		// Released
 		i8 k = SCAN_SET[ret - 0x80];
-		if (!keys[k]) {
+		if (!keys[(u32)k]) {
 			return '\0';
 		}
-		keys[k] = false;
+		keys[(u32)k] = false;
 		return -k;
 	} else {
 		i8 k = SCAN_SET[ret];
-		if (keys[k]) {
+		if (keys[(u32)k]) {
 			return '\0';
 		}
-		keys[k] = true;
+		keys[(u32)k] = true;
 		return k;
 	}
 	return '\0';
 }
 
 i8 translate_code(i8 in) {
-	if (!(keys[LSHIFT_CODE] || keys[RSHIFT_CODE])) {
+	if (!(keys[(u32)LSHIFT_CODE] || keys[(u32)RSHIFT_CODE])) {
 		return in;
 	}
 	if (in >= 'a' && in <= 'z') {
